smasher: porting to GLES 1.1

+++ Demo/Smasher/Tree.hpp
@@ -1,4 +1,14 @@
 
+#include "3dlib/3dlib.h"
+#include "base3d/CVector.hpp"
+#include "base3d/CTexel.hpp"
+#include "base3d/CColor.hpp"
+#include "base3d/CMatrix.hpp"
+#include "load3ds/Load3ds.h"
+
+#include "TexManager.hpp"
+#include "MainFrame/MainFrame.hpp"
+
 #include "Effect.hpp"
 
 #include "ffd.h"
@@ -20,7 +30,7 @@ protected:
         int m_iVertices;
 
 
-        int* m_pFaces;
+        unsigned short* m_pFaces; // TODO: this should be signed?
         int m_iFaces;
 
         int m_iGLTex;
